Fix Log; CLI Parse; Multi-Pkg Support;
I am now using a log.Logger to do all logging. This solves all of my
problems with that :P
A little refactoring of that and others comes with that.

Also proper CLI parsing exists. There is a -u flag but it does nothing.
Multiple packages can also be passed on the command line.